fix(trusty-common): extract_section skips header line instead of failing closed
Code review (code-critic) found a CRITICAL regression in the previous
commit's extract_section hardening: rejecting any header with trailing
text and returning None for the whole section is worse than the leak it
fixed. Annotated headers like "## In Progress (BACKGROUND AGENTS DIE AT
PAUSE — ...)" were this project's own standard hand-written convention
for weeks before pause.rs's native writer existed — not an isolated
incident. Simulated against this project's own 50-file
.trusty-mpm/sessions/*.md archive (07-06 through 07-24, 300 section
extractions): the fail-closed version regressed 64 from present content
to None, and render_session silently omits a None section with no
warning, so a resume digest would misreport "nothing in progress" for
sessions with real unfinished work.

Fix: skip past the header's own line (through its own newline) before
capturing the body, regardless of what follows the header text on that
line. This still fixes the original #3901 leak (trailing annotation text
no longer ends up prepended to the body) without losing content from any
annotated historical header. Re-simulated against the same 50-file
corpus: zero regressions, 64 sections corrected (leaked prefix removed).

- Replaced the fail-closed loop with the simpler skip-the-line approach;
  dropped the "prefer a later well-formed header" search since it's no
  longer meaningful (first occurrence always wins now) and that scenario
  doesn't occur anywhere in the real corpus.
- Rewrote extract_section's regression tests: the trailing-annotation
  test now asserts the annotation does NOT leak into the body (not that
  the section is absent), added a corpus-representative test covering a
  parenthetical annotation, an em-dash annotation, and a clean header —
  all pulled from real files in the archive.
- Completed the `Test:` doc pointer to list all five tests covering
  extract_section.
- Split the growing test module into a sibling session_finder_tests.rs
  (existing #[path] convention used elsewhere in this crate) to keep
  session_finder.rs under the 500-SLOC production cap.

Still closes #3901.
