+++ crates/trusty-common/CHANGELOG.md
@@ -15,11 +15,23 @@ Format follows [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   as a match for `"Next Steps"`, silently prepending the trailing annotation
   into the parsed section body — a real corruption reproduced from this
   project's own `.trusty-mpm/sessions/session-20260721-020826.md`. The
-  matcher now requires only whitespace between the header text and the line
-  break (or end-of-file) before accepting a match, and keeps scanning past a
-  malformed candidate in case a later, well-formed occurrence of the same
-  header exists. Regression tests added using the real corrupted header
-  shape as the fixture.
+  matcher now skips past the entire header LINE (through its own newline)
+  before capturing the body, regardless of what follows the header text on
+  that line, so the trailing annotation never leaks in. An earlier draft of
+  this fix instead rejected any header with trailing text outright
+  (fail-closed, returning `None` for the whole section) — code review caught
+  that this was a worse regression than the bug it fixed: annotated headers
+  (`## In Progress (BACKGROUND AGENTS DIE AT PAUSE — …)`, `## Next Steps
+  (RESUME PLAYBOOK)`) were this project's own standard hand-written
+  convention for weeks before the native `pause.rs` writer existed, not an
+  isolated incident, and `render_session` silently omits a `None` section
+  with no warning. Simulated against this project's own 50-file
+  `.trusty-mpm/sessions/*.md` archive (07-06 through 07-24, 300 section
+  extractions): the fail-closed draft regressed 64 from present-content to
+  `None`; the shipped line-skip fix regresses zero and additionally corrects
+  64 sections that previously carried leaked annotation-text prefixes.
+  Regression tests added, including a representative sample of the corpus's
+  annotated-header styles.
 
 ---
 ## [0.26.1] — 2026-07-24
