+++ crates/trusty-common/src/catchup/session_finder.rs
@@ -309,46 +309,49 @@ fn parse_trusty_mpm_session(path: &Path) -> anyhow::Result<PausedSession> {
 /// Extract the content of a `## <header>` section from markdown text.
 ///
 /// Why: the trusty-mpm session format uses level-2 headers as section delimiters.
-/// What: returns the trimmed content between `## <header>` and the next `## `
-/// or end-of-file. Returns `None` when the section is absent.
+/// What: returns the trimmed content between the end of the `## <header>`
+/// LINE and the next `## ` header or end-of-file. Returns `None` when the
+/// section is absent or empty.
 ///
-/// Fails closed on a malformed header: `text.find` alone would treat
-/// `## Next Steps (all Bob's call — none required)` as a match for header
-/// `"Next Steps"` and silently prepend the trailing `(all Bob's call — none
-/// required)` text into the parsed body (a real corruption seen in a
-/// hand-written snapshot, see `extract_section_rejects_header_with_trailing_text`).
-/// A match is only accepted when nothing but whitespace follows the header
-/// text before the line break (or end-of-file) — i.e. the header line reads
-/// exactly `## <header>`, optionally with trailing spaces. Malformed
-/// candidates are skipped in favor of a later well-formed occurrence, if any.
+/// Trailing text on the header line (e.g. a hand-written `## Next Steps
+/// (all Bob's call — none required)`) is skipped along with the rest of that
+/// line rather than being rejected outright. An earlier version of this
+/// function rejected any header with trailing text and returned `None` for
+/// the whole section — that is a WORSE bug than the one it was meant to fix:
+/// annotated headers (`## In Progress (BACKGROUND AGENTS DIE AT PAUSE — …)`,
+/// `## Next Steps (RESUME PLAYBOOK)`, `## Completed (this leg)`) were this
+/// project's own standard hand-written convention for weeks before
+/// `pause.rs`'s native writer existed, not an isolated one-off. Simulating
+/// the fail-closed version against this project's own 50-file
+/// `.trusty-mpm/sessions/*.md` archive (07-06 through 07-24) regressed 64 of
+/// 300 section extractions from present content to `None` — and
+/// `render_session` silently *omits* a `None` section with no warning, so a
+/// resume digest would misreport "nothing in progress" for sessions that
+/// recorded substantial unfinished work. Skipping just the header's own
+/// line — regardless of its trailer — fixes the original leak (the trailing
+/// annotation text no longer ends up prepended to the body) without losing
+/// content from any annotated historical header.
 /// Test: `extract_section_finds_content`,
-/// `extract_section_rejects_header_with_trailing_text`.
+/// `extract_section_strips_trailing_header_annotation_from_body`,
+/// `extract_section_allows_trailing_whitespace_on_header_line`,
+/// `extract_section_first_occurrence_wins_when_header_repeated`,
+/// `extract_section_handles_representative_corpus_header_styles`.
 fn extract_section(text: &str, header: &str) -> Option<String> {
     let needle = format!("## {header}");
-    let mut search_from = 0;
-    while let Some(rel_start) = text[search_from..].find(&needle) {
-        let start = search_from + rel_start;
-        let after_needle = start + needle.len();
-        let rest = &text[after_needle..];
-        let line_end = rest.find('\n').unwrap_or(rest.len());
-        let header_trailer = &rest[..line_end];
-        if !header_trailer.trim().is_empty() {
-            // Header line has extra text after it (e.g. a hand-written
-            // "## Next Steps (all Bob's call — none required)") — this is
-            // not the delimiter we're looking for. Keep scanning past it
-            // rather than absorbing the trailing text into the body.
-            search_from = after_needle;
-            continue;
-        }
-        let end = rest.find("\n## ").unwrap_or(rest.len());
-        let section = rest[..end].trim().to_owned();
-        return if section.is_empty() {
-            None
-        } else {
-            Some(section)
-        };
+    let start = text.find(&needle)?;
+    let after_needle = start + needle.len();
+    let rest = &text[after_needle..];
+    // Skip past the rest of the header LINE (through its own newline, if
+    // any) so trailing annotation text never leaks into the captured body.
+    let line_end = rest.find('\n').map(|i| i + 1).unwrap_or(rest.len());
+    let body = &rest[line_end..];
+    let end = body.find("\n## ").unwrap_or(body.len());
+    let section = body[..end].trim().to_owned();
+    if section.is_empty() {
+        None
+    } else {
+        Some(section)
     }
-    None
 }
 
 /// Parse a timestamp from the `YYYYMMDD-HHMMSS` portion of a session filename.
@@ -379,297 +382,5 @@ fn parse_filename_timestamp(stem: &str) -> Option<DateTime<Utc>> {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::catchup::mpm_session::ClaudeMpmSession;
-    use std::fs;
-    use tempfile::TempDir;
-
-    fn write_file(dir: &Path, name: &str, content: &str) -> PathBuf {
-        let p = dir.join(name);
-        fs::write(&p, content).unwrap();
-        p
-    }
-
-    #[test]
-    fn parse_filename_timestamp_roundtrip() {
-        let ts = parse_filename_timestamp("20260627-142030");
-        assert!(ts.is_some());
-        let ts = ts.unwrap();
-        assert_eq!(ts.format("%Y-%m-%d").to_string(), "2026-06-27");
-    }
-
-    #[test]
-    fn parse_filename_timestamp_rejects_short() {
-        assert!(parse_filename_timestamp("2026062").is_none());
-        assert!(parse_filename_timestamp("").is_none());
-    }
-
-    #[test]
-    fn extract_section_finds_content() {
-        let md = "# Title\n\n## Summary\nDid lots of work.\n\n## Next Steps\nFix tests.";
-        assert_eq!(
-            extract_section(md, "Summary").as_deref(),
-            Some("Did lots of work.")
-        );
-        assert_eq!(
-            extract_section(md, "Next Steps").as_deref(),
-            Some("Fix tests.")
-        );
-        assert!(extract_section(md, "Missing").is_none());
-    }
-
-    /// Regression test for issue #3901: a hand-written snapshot header with
-    /// trailing text (`## Next Steps (all Bob's call — none required)`) must
-    /// NOT be treated as a match for the `"Next Steps"` section — the old
-    /// substring-`find` implementation absorbed the trailing `(all Bob's
-    /// call — none required)` into the parsed body. Fixture text is the
-    /// actual corrupted header shape from
-    /// `.trusty-mpm/sessions/session-20260721-020826.md` in this project.
-    #[test]
-    fn extract_section_rejects_header_with_trailing_text() {
-        let md = "## In Progress\n\nNOTHING.\n\n\
-                   ## Next Steps (all Bob's call — none required)\n\n\
-                   - Bob's tested one-liner (handed over, not yet run).\n\
-                   - Installer NOT VERIFIED in isolation.\n";
-        // No well-formed "## Next Steps" header exists, so the section is
-        // absent rather than corrupted with the trailing header text.
-        assert!(extract_section(md, "Next Steps").is_none());
-        // The well-formed sibling header still parses correctly.
-        assert_eq!(
-            extract_section(md, "In Progress").as_deref(),
-            Some("NOTHING.")
-        );
-    }
-
-    /// A malformed header earlier in the document must not shadow a later,
-    /// well-formed occurrence of the same header.
-    #[test]
-    fn extract_section_finds_later_well_formed_header_after_malformed_one() {
-        let md = "## Next Steps (draft, ignore)\nstale draft text\n\n\
-                   ## Next Steps\nReal next steps.\n";
-        assert_eq!(
-            extract_section(md, "Next Steps").as_deref(),
-            Some("Real next steps.")
-        );
-    }
-
-    /// Trailing whitespace after the header (no visible extra text) is still
-    /// a well-formed header and must parse normally.
-    #[test]
-    fn extract_section_allows_trailing_whitespace_on_header_line() {
-        let md = "## Next Steps   \nDo the thing.\n";
-        assert_eq!(
-            extract_section(md, "Next Steps").as_deref(),
-            Some("Do the thing.")
-        );
-    }
-
-    #[test]
-    fn find_merges_both_formats() {
-        let tmp = TempDir::new().unwrap();
-        let project = tmp.path();
-
-        // Create trusty-mpm session file.
-        let tm_dir = project.join(".trusty-mpm").join("sessions");
-        fs::create_dir_all(&tm_dir).unwrap();
-        write_file(
-            &tm_dir,
-            "session-20260627-100000.md",
-            "## Summary\nDone something.\n## Git Context\nbranch: main",
-        );
-
-        // Create claude-mpm session file.
-        let cm_dir = project.join(".claude-mpm").join("sessions");
-        fs::create_dir_all(&cm_dir).unwrap();
-        write_file(
-            &cm_dir,
-            "session-20260626-090000.json",
-            r#"{"session_id":"cm1","paused_at":"2026-06-26T09:00:00Z"}"#,
-        );
-
-        let sessions = find_paused_sessions(project).unwrap();
-        assert_eq!(sessions.len(), 2);
-        assert!(
-            matches!(sessions[0], PausedSession::TrustyMpm { .. }),
-            "newer trusty-mpm session should be first"
-        );
-        assert!(
-            matches!(sessions[1], PausedSession::ClaudeMpm { .. }),
-            "older claude-mpm session should be second"
-        );
-    }
-
-    #[test]
-    fn find_orders_newest_first() {
-        let tmp = TempDir::new().unwrap();
-        let cm_dir = tmp.path().join(".claude-mpm").join("sessions");
-        fs::create_dir_all(&cm_dir).unwrap();
-        write_file(
-            &cm_dir,
-            "session-20260625-080000.json",
-            r#"{"session_id":"old","paused_at":"2026-06-25T08:00:00Z"}"#,
-        );
-        write_file(
-            &cm_dir,
-            "session-20260627-100000.json",
-            r#"{"session_id":"new","paused_at":"2026-06-27T10:00:00Z"}"#,
-        );
-
-        let sessions = find_paused_sessions(tmp.path()).unwrap();
-        assert_eq!(sessions.len(), 2);
-        let first_key = sessions[0].sort_key().unwrap();
-        let second_key = sessions[1].sort_key().unwrap();
-        assert!(first_key > second_key, "newest should be first");
-    }
-
-    #[test]
-    fn render_contains_digest_not_conversation() {
-        let session = ClaudeMpmSession {
-            session_id: "test-123".to_string(),
-            paused_at: Some("2026-06-27T10:00:00Z".to_string()),
-            resume_instructions: Some("Resume from step 3".to_string()),
-            important_reminders: Some(vec!["Don't break prod".to_string()]),
-            git_context: Some("branch: main".to_string()),
-            ..Default::default()
-        };
-        let sessions = vec![PausedSession::ClaudeMpm { session }];
-        let output = render_resume_context(&sessions);
-        assert!(
-            output.contains("Resume from step 3"),
-            "resume instructions should be present"
-        );
-        assert!(
-            output.contains("Don't break prod"),
-            "reminders should be present"
-        );
-        assert!(
-            output.contains("branch: main"),
-            "git context should be present"
-        );
-        assert!(
-            !output.contains("conversation"),
-            "conversation must NOT appear in rendered output"
-        );
-    }
-
-    #[test]
-    fn parse_extracts_tmux_window() {
-        let tmp = TempDir::new().unwrap();
-        let p = write_file(
-            tmp.path(),
-            "session-20260627-100000.md",
-            "## Summary\nWork.\n\n## Tmux Window\nmain:2:@7\n\n## Git Context\nbranch: main",
-        );
-        let session = parse_trusty_mpm_session(&p).unwrap();
-        match session {
-            PausedSession::TrustyMpm { tmux_window, .. } => {
-                assert_eq!(tmux_window.as_deref(), Some("main:2:@7"));
-            }
-            _ => panic!("expected TrustyMpm variant"),
-        }
-    }
-
-    #[test]
-    fn parse_missing_tmux_window_is_none() {
-        let tmp = TempDir::new().unwrap();
-        let p = write_file(
-            tmp.path(),
-            "session-20260627-100000.md",
-            "## Summary\nWork.\n\n## Git Context\nbranch: main",
-        );
-        let session = parse_trusty_mpm_session(&p).unwrap();
-        match session {
-            PausedSession::TrustyMpm { tmux_window, .. } => {
-                assert!(tmux_window.is_none(), "back-compat: absent section => None");
-            }
-            _ => panic!("expected TrustyMpm variant"),
-        }
-    }
-
-    #[test]
-    fn render_tmux_window_present_and_omitted() {
-        // Present → line renders.
-        let with = PausedSession::TrustyMpm {
-            path: PathBuf::from("/tmp/session-20260627-100000.md"),
-            paused_at: None,
-            summary: "Work.".to_string(),
-            git_context: None,
-            in_progress: None,
-            next_steps: None,
-            tmux_window: Some("main:2:@7".to_string()),
-        };
-        let output = render_resume_context(std::slice::from_ref(&with));
-        assert!(
-            output.contains("**Tmux Window:** main:2:@7"),
-            "recorded window should render"
-        );
-
-        // None → line omitted.
-        let without = PausedSession::TrustyMpm {
-            path: PathBuf::from("/tmp/session-20260627-100000.md"),
-            paused_at: None,
-            summary: "Work.".to_string(),
-            git_context: None,
-            in_progress: None,
-            next_steps: None,
-            tmux_window: None,
-        };
-        let output = render_resume_context(std::slice::from_ref(&without));
-        assert!(
-            !output.contains("Tmux Window"),
-            "absent window must not render a line"
-        );
-    }
-
-    #[test]
-    fn latest_snapshot_prefers_session_log() {
-        use crate::catchup::session_log::{SessionLogEntry, append_entry};
-        let tmp = TempDir::new().unwrap();
-        let sdir = tmp.path().join(".trusty-mpm").join("sessions");
-        fs::create_dir_all(&sdir).unwrap();
-
-        // Two sessions interleave; each snapshot file exists on disk.
-        write_file(&sdir, "session-A.md", "## Summary\nS1 work");
-        write_file(&sdir, "session-B.md", "## Summary\nS2 work");
-        let mk = |id: &str, snap: &str, ts: &str| SessionLogEntry {
-            session_id: id.to_string(),
-            event: "pause".to_string(),
-            snapshot: snap.to_string(),
-            timestamp: ts.to_string(),
-        };
-        append_entry(&sdir, &mk("s1", "session-A.md", "t1")).unwrap();
-        append_entry(&sdir, &mk("s2", "session-B.md", "t2")).unwrap();
-
-        // s1 resumes its own snapshot even though s2 paused last.
-        let got = latest_trusty_mpm_snapshot(tmp.path(), Some("s1")).unwrap();
-        assert_eq!(got.file_name().unwrap(), "session-A.md");
-        // No id → latest overall (s2's).
-        let got = latest_trusty_mpm_snapshot(tmp.path(), None).unwrap();
-        assert_eq!(got.file_name().unwrap(), "session-B.md");
-    }
-
-    #[test]
-    fn latest_snapshot_falls_back() {
-        let tmp = TempDir::new().unwrap();
-        let sdir = tmp.path().join(".trusty-mpm").join("sessions");
-        fs::create_dir_all(&sdir).unwrap();
-        // No log at all → mtime scan of session-*.md.
-        write_file(&sdir, "session-20260101-000000.md", "## Summary\nold");
-        std::thread::sleep(std::time::Duration::from_millis(10));
-        write_file(&sdir, "session-20260202-000000.md", "## Summary\nnew");
-        let got = latest_trusty_mpm_snapshot(tmp.path(), Some("unknown-id")).unwrap();
-        assert_eq!(got.file_name().unwrap(), "session-20260202-000000.md");
-        // Missing project dir → None.
-        assert!(latest_trusty_mpm_snapshot(&tmp.path().join("nope"), None).is_none());
-    }
-
-    #[test]
-    fn render_empty_returns_no_sessions_message() {
-        let output = render_resume_context(&[]);
-        assert!(
-            output.contains("No paused sessions"),
-            "empty renders a notice"
-        );
-    }
-}
+#[path = "session_finder_tests.rs"]
+mod tests;
